backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
torch==2.1.1
torchvision==0.16.1
pytorch-lightning==2.1.2
scikit-learn==1.3.2
pandas==2.1.3
pillow==10.1.0
numpy==1.26.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
